requirements.txt
certifi==2021.5.30
chardet==4.0.0
idna==2.10
python-dotenv==0.18.0
requests==2.25.1
tqdm==4.61.1
urllib3==1.26.6
